+++ app/src/main/java/at/jku/yourmemorylane/db/entities/Media.kt
@@ -6,9 +6,9 @@ import androidx.room.PrimaryKey
 
 @Entity
 data class Media(
-    @ColumnInfo(name = "memoryId") val memoryId: Int,
+    @ColumnInfo(name = "memoryId") val memoryId: Long,
     @ColumnInfo(name = "type") val type: String,
     @ColumnInfo(name = "path") val path: String
 ){
-    @PrimaryKey(autoGenerate = true) var id: Int =0
+    @PrimaryKey(autoGenerate = true) var id: Long = 0
 }
